% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_varcategory.R
\name{add_varcategory}
\alias{add_varcategory}
\title{Add numeric variable to travel times}
\usage{
add_varcategory(thisnum, names.categ, pov.ind.codes)
}
\arguments{
\item{pov.ind.codes}{}
}
\description{
Add numeric variable to travel times
}
