% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_topsis.R
\name{calc_topsis}
\alias{calc_topsis}
\title{Calculate vulnerability}
\usage{
calc_topsis(
  indicator.name.sens,
  indicator.name.exp,
  indicator.name.adap,
  cost.data.comms,
  file.ind
)
}
\arguments{
\item{file.ind}{}
}
\description{
Calculate vulnerability
}
\examples{
calc_topsis(indicator.name.sens="sensitivity", indicator.name.exp= "exposure_historical", indicator.name.adap= "adaptive_capacity", cost.data.comms, file.ind = "historical")
}
