% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bienpesca.R
\name{clean_bienpesca}
\alias{clean_bienpesca}
\title{Clean Bienpesca data}
\usage{
clean_bienpesca(this.data)
}
\arguments{
\item{this.data}{}
}
\value{
clean.data
}
\description{
Clean Bienpesca data
}
