% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_seedfiles.R
\name{clean_seedfiles}
\alias{clean_seedfiles}
\title{clean subsidy files to support seed purchase for aquaculture}
\usage{
clean_seedfiles(thisfile, inegi.cost.comm)
}
\arguments{
\item{inegi.cost.comm}{coastal communities with inegi data}
}
\value{
new.data
}
\description{
clean subsidy files to support seed purchase for aquaculture
}
