% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raster.R
\name{expo_raster}
\alias{expo_raster}
\title{funcion para exposicion}
\usage{
expo_raster(thisnum, vars.expo, coste.buff, cost.data.pob)
}
\description{
Calcula datos de exposicion y produce graficas espaciales, por estado y por municipio
}
\details{
INPUT: 1) data

OUTPUT: 1) resultados pca, 2) graficos
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
