% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_comms.R
\name{extract_comms}
\alias{extract_comms}
\title{Extract target states and municipalities}
\usage{
extract_comms(this.data)
}
\arguments{
\item{inegi.cost.comm}{}
}
\value{
this.data.selec
}
\description{
Extract target states and municipalities
}
