% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_denue.R
\name{extract_denue}
\alias{extract_denue}
\title{Extract denue data for coastal communities}
\usage{
extract_denue(
  eachdenue,
  denue.inegi,
  denue.inegi.full,
  input.dir,
  inegi.comm.codes
)
}
\arguments{
\item{inegi.comm.codes}{}
}
\description{
Extract denue data for coastal communities
}
