% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract landing locations}
\usage{
extract_locations(
  this.catchfileno,
  catch.files,
  coast.selecc,
  locs.pesca,
  inegi.cost.comm.all
)
}
\arguments{
\item{coast.selecc}{}
}
\description{
Extract landing locations
}
