% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_permits.R
\name{extract_permits}
\alias{extract_permits}
\title{Extract permit information}
\usage{
extract_permits(thispermitfile.no, permit.files, inegi.cost.comm, coast.selecc)
}
\arguments{
\item{coast.selecc}{}
}
\value{
this.permit.inegi
}
\description{
Extract permit with location information to match with catch records
}
