% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raster.R
\name{extract_raster}
\alias{extract_raster}
\title{funcion para exposicion}
\usage{
extract_raster(
  thisnum,
  vars.expo,
  inegi.cost.buffer.50,
  cost.data.pob,
  r.template,
  crs.proj.wgs
)
}
\description{
Calcula datos de exposicion y produce graficas espaciales, por estado y por municipio
}
\details{
INPUT: 1) data

OUTPUT: 1) resultados pca, 2) graficos
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
