% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_inegi.R
\name{fix_inegi}
\alias{fix_inegi}
\title{Standarize INEGI table for communities}
\usage{
fix_inegi(inegi.table)
}
\arguments{
\item{inegi.table}{}
}
\value{
inegi.comms
}
\description{
Standarize INEGI table for communities
}
