% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_data.R
\name{get_community_data}
\alias{get_community_data}
\title{get community data}
\usage{
get_community data()
}
\arguments{
\item{crs.proj.utm, crs.proj.wgs, }{atlantic.shp, pacific.shp, est.2019, inegi.2020}
}
\value{
inegi.cost.wgs
}
\description{
gets location data
}
\details{
INPUT: 1) data

OUTPUT: 1) locations
}
\keyword{function}
\keyword{inegi}
