% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_data.R
\name{get_comp_data}
\alias{get_comp_data}
\title{Combine composition total scores}
\usage{
get_comp_data(
  indicator.name.sens,
  indicator.name.exp,
  indicator.name.adap,
  cost.data.comms,
  file.ind
)
}
\arguments{
\item{cost.data.comms}{}
}
\description{
Combine composition total scores
}
