% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rnpa_catch.R
\name{get_rnpa_catch}
\alias{get_rnpa_catch}
\title{Extrac RNPA codes from catch files}
\usage{
get_rnpa_catch(this.catchfileno, catch.files, coast.selecc)
}
\arguments{
\item{catch.files}{catch files}

\item{coast.selecc}{selected coastal states}
}
\value{
catch.file.rnp
}
\description{
Extrac RNPA codes from catch files
}
