% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_worldclim.R
\name{get_worldclim}
\alias{get_worldclim}
\title{Get data from WorldClim}
\usage{
get_worldclim(
  eachscenario,
  scenario.names,
  file.names,
  var.names,
  layer.nums,
  input.dir,
  r.template
)
}
\arguments{
\item{input.dir}{}
}
\description{
Get terrestrial historical and future layers from https://www.worldclim.org/data/cmip6/cmip6_clim30s.html
}
