% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_inegi.R
\name{local_inegi}
\alias{local_inegi}
\title{Query INEGI location catalog}
\usage{
local_inegi(this.loc, ent.codes)
}
\value{
data.locs
}
\description{
Obtener localidades del catalogo de INEGI usando el servicio JSON

https://www.inegi.org.mx/servicios/catalogounico.html
}
\author{
Hem Nalini Morzaria Luna
}
