% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_code.R
\name{make_code}
\alias{make_code}
\title{Add code based on locality, municipality and state codes}
\usage{
make_code(this.tibble)
}
\arguments{
\item{this.tibble}{}
}
\value{
code.tibble
}
\description{
Add code based on locality, municipality and state codes
}
