% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Mapping coastal communities}
\usage{
make_map()
}
\arguments{
\item{estados.shp, }{inegi.cost.wgs, thistitle, thisoutput, ents.selecc}
}
\value{
map.locs
}
\description{
Functions to make map
}
\details{
Make map of coastal communities 30 km from coast
}
\keyword{function}
\keyword{map}
