% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mangroveclasses.R
\docType{data}
\name{mangroveclasses}
\alias{mangroveclasses}
\title{DATASET_TITLE}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{\code{CAMBIOS}}{character COLUMN_DESCRIPTION}
\item{\code{new_value}}{character COLUMN_DESCRIPTION}
}
}
\usage{
mangroveclasses
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
