% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vul.R
\name{map_vul}
\alias{map_vul}
\title{Title}
\usage{
map_vul(
  this.data,
  crs.proj.wgs,
  crs.proj.utm,
  plot.title = "Distribución espacial de la vulnerabilidad actual",
  scenario.nom = "historical",
  mun.cost
)
}
\arguments{
\item{mun.cost}{}
}
\description{
Title
}
