% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Merge indicator files}
\usage{
merge_files(indicator.list, cost.data.pob, indicator.name)
}
\arguments{
\item{cost.data.pob}{}
}
\value{
ind.pob.sc
}
\description{
Merge indicator files
}
