% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_factor.R
\name{pca_factor}
\alias{pca_factor}
\title{PCA of indicators}
\usage{
pca_factor(this.data, indicator.name)
}
\arguments{
\item{indicator.name}{}
}
\description{
PCA of indicators
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
