% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change.R
\name{plot_change}
\alias{plot_change}
\title{Plot component scores}
\usage{
plot_change(
  this.data.hist,
  this.data.ssp126,
  this.data.ssp585,
  ent.cols,
  ent.ord,
  title.txt
)
}
\arguments{
\item{title.txt}{}
}
\description{
Plot component scores
}
\examples{
plot_change(this.data.hist=tot.scores.comp.hist, this.data.ssp126=tot.scores.comp.ssp126, this.data.ssp585=tot.scores.comp.ssp585, ent.cols, ent.ord, title.txt="Cambio en exposición relativo a condiciones actuales")
}
