% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comps.R
\name{plot_comps}
\alias{plot_comps}
\title{Plot component scores}
\usage{
plot_comps(this.data, cost.data.comms, scenario.name, cols, ent.ord, title.txt)
}
\arguments{
\item{title.txt}{}
}
\description{
Plot component scores
}
\examples{
this.data.hist = tot.scores.comp.hist, this.data.ssp126 = tot.scores.comp.ssp126, this.data.ssp585 = tot.scores.comp.ssp585
}
