% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_var_munOLD.R
\name{scale_var_munold}
\alias{scale_var_munold}
\title{Scale variable}
\usage{
scale_var_munold(this.data, indicator.name)
}
\arguments{
\item{sspscenario}{}
}
\value{
num.scale data with scaled total
}
\description{
Scale variable
}
\examples{
scaled.exp <- scale_var(this.data)
}
