% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_files.R
\name{select_files}
\alias{select_files}
\title{Merge indicator files}
\usage{
select_files(this.ind.num, indicator.list, cost.data.pob, indicator.name)
}
\arguments{
\item{cost.data.pob}{}
}
\value{
ind.pob.sc
}
\description{
Merge indicator files
}
