% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-temprange.R
\docType{data}
\name{temprange}
\alias{temprange}
\title{DATASET_TITLE}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{\code{TEM_C}}{character COLUMN_DESCRIPTION}
\item{\code{new_value}}{character COLUMN_DESCRIPTION}
}
}
\usage{
temprange
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
