% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_score.R
\name{get_comp_score}
\alias{get_comp_score}
\title{Get total component score}
\usage{
get_comp_score(this.ent.data, this.mun.data, indicator.name)
}
\arguments{
\item{indicator.name}{}
}
\description{
Get total component score
}
